Refactor database connection management
Database connection is now a property of the SgrMnr class which is opened when the form is loaded and closed during destruction.

The database is default created and initialized whenever the form loads (if it doesn't already exist)

A settings table is added with basic name/value pairs in text.  The only setting for now is ratio.

The tab stops were updated to make it flow more intuitively when using tabs.

The refresh button was hooked up.